refactor sklearnex get_namespace usage
