+++ onedal/utils/_array_api.py
@@ -59,3 +59,26 @@ def _from_dlpack(data, xp, *args, **kwargs):
 def _is_numpy_namespace(xp):
     """Return True if xp is backed by NumPy."""
     return xp.__name__ in {"numpy", "array_api_compat.numpy", "numpy.array_api"}
+
+
+def _get_sycl_namespace(*arrays):
+    """Get namespace of sycl arrays."""
+
+    # sycl support designed to work regardless of array_api_dispatch sklearn global value
+    sycl_type = {type(x): x for x in arrays if hasattr(x, "__sycl_usm_array_interface__")}
+
+    if len(sycl_type) > 1:
+        raise ValueError(f"Multiple SYCL types for array inputs: {sycl_type}")
+
+    if sycl_type:
+        (X,) = sycl_type.values()
+
+        if hasattr(X, "__array_namespace__"):
+            return sycl_type, X.__array_namespace__(), True
+        elif dpnp_available and isinstance(X, dpnp.ndarray):
+            # convert it to dpctl.tensor with namespace.
+            return sycl_type, dpnp, False
+        else:
+            raise ValueError(f"SYCL type not recognized: {sycl_type}")
+
+    return sycl_type, None, False
