+++ sklearnex/utils/_array_api.py
@@ -17,16 +17,14 @@
 import numpy as np
 
 from daal4py.sklearn._utils import sklearn_check_version
-
-from .._device_offload import dpnp_available
+from onedal.utils._array_api import _get_sycl_namespace
 
 if sklearn_check_version("1.2"):
     from sklearn.utils._array_api import get_namespace as sklearn_get_namespace
 
-if dpnp_available:
-    import dpnp
 
-
+# TODO:
+# update it for supported versions of scikit-learn.
 def get_namespace(*arrays):
     """Get namespace of arrays.
 
@@ -74,23 +72,10 @@ def get_namespace(*arrays):
         True of the arrays are containers that implement the Array API spec.
     """
 
-    # sycl support designed to work regardless of array_api_dispatch sklearn global value
-    sycl_type = {type(x): x for x in arrays if hasattr(x, "__sycl_usm_array_interface__")}
-
-    if len(sycl_type) > 1:
-        raise ValueError(f"Multiple SYCL types for array inputs: {sycl_type}")
+    sycl_type, xp, is_array_api_compliant = _get_sycl_namespace(*arrays)
 
     if sycl_type:
-
-        (X,) = sycl_type.values()
-
-        if hasattr(X, "__array_namespace__"):
-            return X.__array_namespace__(), True
-        elif dpnp_available and isinstance(X, dpnp.ndarray):
-            return dpnp, False
-        else:
-            raise ValueError(f"SYCL type not recognized: {sycl_type}")
-
+        return xp, is_array_api_compliant
     elif sklearn_check_version("1.2"):
         return sklearn_get_namespace(*arrays)
     else:
